+++ apps-rendering/src/client/newsletterSignupForm.ts
